# Speed performance optimized by 30 times