--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "chrono": "cpp",
+        "cmath": "cpp"
+    }
+}