--- a/main.cpp
+++ b/main.cpp
@@ -4,6 +4,57 @@
 #include <chrono>
 #include <cmath>
 
+#include <mmintrin.h> //mmx
+#include <xmmintrin.h> //sse
+#include <emmintrin.h> //sse2
+#include <pmmintrin.h> //sse3
+
+#pragma GCC target("avx")
+#pragma GCC optimize(3)
+#pragma GCC optimize("Ofast")
+#pragma GCC optimize("inline")
+#pragma GCC optimize("-fgcse")
+#pragma GCC optimize("-fgcse-lm")
+#pragma GCC optimize("-fipa-sra")
+#pragma GCC optimize("-ftree-pre")
+#pragma GCC optimize("-ftree-vrp")
+#pragma GCC optimize("-fpeephole2")
+#pragma GCC optimize("-ffast-math")
+#pragma GCC optimize("-fsched-spec")
+#pragma GCC optimize("unroll-loops")
+#pragma GCC optimize("-falign-jumps")
+#pragma GCC optimize("-falign-loops")
+#pragma GCC optimize("-falign-labels")
+#pragma GCC optimize("-fdevirtualize")
+#pragma GCC optimize("-fcaller-saves")
+#pragma GCC optimize("-fcrossjumping")
+#pragma GCC optimize("-fthread-jumps")
+#pragma GCC optimize("-funroll-loops")
+#pragma GCC optimize("-freorder-blocks")
+#pragma GCC optimize("-fschedule-insns")
+#pragma GCC optimize("inline-functions")
+#pragma GCC optimize("-ftree-tail-merge")
+#pragma GCC optimize("-fschedule-insns2")
+#pragma GCC optimize("-fstrict-aliasing")
+#pragma GCC optimize("-falign-functions")
+#pragma GCC optimize("-fcse-follow-jumps")
+#pragma GCC optimize("-fsched-interblock")
+#pragma GCC optimize("-fpartial-inlining")
+#pragma GCC optimize("no-stack-protector")
+#pragma GCC optimize("-freorder-functions")
+#pragma GCC optimize("-findirect-inlining")
+#pragma GCC optimize("-fhoist-adjacent-loads")
+#pragma GCC optimize("-frerun-cse-after-loop")
+#pragma GCC optimize("inline-small-functions")
+#pragma GCC optimize("-finline-small-functions")
+#pragma GCC optimize("-ftree-switch-conversion")
+#pragma GCC optimize("-foptimize-sibling-calls")
+#pragma GCC optimize("-fexpensive-optimizations")
+#pragma GCC optimize("inline-functions-called-once")
+#pragma GCC optimize("-fdelete-null-pointer-checks")
+#pragma G++ ivdep
+#pragma G++ unroll 4
+
 float frand() {
     return (float)rand() / RAND_MAX * 2 - 1;
 }
@@ -11,56 +62,58 @@ float frand() {
 struct Star {
     float px, py, pz;
     float vx, vy, vz;
-    float mass;
-};
-
-std::vector<Star> stars;
+    float mass, mass1;
+    Star(float _px,float _py,float _pz,float _vx,float _vy,float _vz,float _mass):
+        px(_px),py(_py),pz(_pz),vx(_vx),vy(_vy),vz(_vz),mass(_mass){}
+    Star() {}
+}stars[48];
 
 void init() {
-    for (int i = 0; i < 48; i++) {
-        stars.push_back({
-            frand(), frand(), frand(),
-            frand(), frand(), frand(),
-            frand() + 1,
-        });
+    for (uint32_t i = 0; i < 48; i++) {
+        stars[i]= Star( frand(), frand(), frand(), frand(), frand(), frand(), frand() + 1);
     }
 }
 
-float G = 0.001;
-float eps = 0.001;
-float dt = 0.01;
+constexpr float G = 0.001;
+constexpr float eps = 0.001;
+constexpr float dt = 0.01;
 
 void step() {
-    for (auto &star: stars) {
-        for (auto &other: stars) {
-            float dx = other.px - star.px;
-            float dy = other.py - star.py;
-            float dz = other.pz - star.pz;
-            float d2 = dx * dx + dy * dy + dz * dz + eps * eps;
+    for (size_t i=0;i<(uint32_t)48;++i) {
+        for (size_t j=0;j<(uint32_t)48;++j) {
+            #pragma opm simd
+            float dx = stars[j].px - stars[i].px;
+            float dy = stars[j].py - stars[i].py;
+            float dz = stars[j].pz - stars[i].pz;
+            float d2 = dx * dx + dy * dy + dz * dz + (eps * eps);
             d2 *= sqrt(d2);
-            star.vx += dx * other.mass * G * dt / d2;
-            star.vy += dy * other.mass * G * dt / d2;
-            star.vz += dz * other.mass * G * dt / d2;
+            d2 = stars[j].mass * G * dt / d2;
+            stars[i].vx += dx * d2;
+            stars[i].vy += dy * d2;
+            stars[i].vz += dz * d2;
         }
     }
-    for (auto &star: stars) {
-        star.px += star.vx * dt;
-        star.py += star.vy * dt;
-        star.pz += star.vz * dt;
+    for(size_t i=0;i<48; ++i){
+        #pragma opm simd
+        stars[i].px += stars[i].vx * dt;
+        stars[i].py += stars[i].vy * dt;
+        stars[i].pz += stars[i].vz * dt;
     }
 }
 
 float calc() {
     float energy = 0;
-    for (auto &star: stars) {
-        float v2 = star.vx * star.vx + star.vy * star.vy + star.vz * star.vz;
-        energy += star.mass * v2 / 2;
-        for (auto &other: stars) {
-            float dx = other.px - star.px;
-            float dy = other.py - star.py;
-            float dz = other.pz - star.pz;
-            float d2 = dx * dx + dy * dy + dz * dz + eps * eps;
-            energy -= other.mass * star.mass * G / sqrt(d2) / 2;
+    for (size_t i=0;i<48;++i) {
+        #pragma opm simd
+        float v2 = stars[i].vx * stars[i].vx + stars[i].vy * stars[i].vy + stars[i].vz * stars[i].vz;
+        energy += stars[i].mass * v2 * 0.5;
+        for (size_t j=0;j<48;++j) {
+            #pragma opm simd
+            float dx = stars[j].px - stars[i].px;
+            float dy = stars[j].py - stars[i].py;
+            float dz = stars[j].pz - stars[i].pz;
+            float d2 =  (dx * dx + dy * dy + dz * dz + (eps * eps));
+            energy -= stars[j].mass * stars[i].mass * 0.0005 / sqrt(d2);
         }
     }
     return energy;