--- a/rbq.bat
+++ b/rbq.bat
@@ -1,0 +1,6 @@
+g++ -o main.exe main.cpp  -ffast-math -march=native -msse4.1 -O2 -fopenmp -O3 -std=c++17
+main.exe
+main.exe
+main.exe
+main.exe
+main.exe