@WangYuHang-cmd 2023-03-16
> #pragma opm simd 错了，是omp 无法顺畅的大口呼吸，是活着的最好证明
> […](#)
> ---Original--- From: ***@***.***&gt; Date: Wed, Mar 15, 2023 17:23 PM To: ***@***.***&gt;; Cc: ***@***.***&gt;; Subject: [parallel101/hw04] Speed performance optimized by 30 times (PR #39) 初始数据: Initial energy: -8.571526 Final energy: -8.511777 Time elapsed: 6646 ms 编译指令加入O3优化： Initial energy: -8.571526 Final energy: -8.511777 Time elapsed: 1737 ms 将结构体OOP改成DOP Initial energy: -8.571526 Final energy: -8.511777 Time elapsed: 1734 ms 加入编译指令 #pragma GCC ivdep #pragma GCC unroll 4 Initial energy: -8.571302 Final energy: -8.511518 Time elapsed: 1587 ms 加上暴力火车头： Initial energy: -8.571527 Final energy: -8.511723 Time elapsed: 1175 ms 加入编译指令： -ffast-math -march=native Initial energy: -8.571527 Final energy: -8.511747 Time elapsed: 210 ms You can view, comment on, or merge this pull request online at: &nbsp; #39 Commit Summary [abefa6b](https://github.com/parallel101/hw04/commit/abefa6b75b7ba56c0a5d1d510b125f878181c3ee) Modified fast [3e87fbf](https://github.com/parallel101/hw04/commit/3e87fbf5f685f16a59b5490a65d3b878c8d4bb75) Modified fast File Changes (8 files) A .vscode/settings.json (6) M CMakeLists.txt (8) M README.md (44) A initial.cpp (88) A initial.exe (0) M main.cpp (127) A main.exe (0) A rbq.bat (6) Patch Links: https://github.com/parallel101/hw04/pull/39.patch https://github.com/parallel101/hw04/pull/39.diff — Reply to this email directly, view it on GitHub, or unsubscribe. You are receiving this because you are subscribed to this thread.Message ID: ***@***.***&gt;

尴尬....😂